Added observe_self and try_observe_self to EntityCommands
`observe_self` inserts an `Observer` observing the associated entity of the `EntityCommands` while providing the same method chaining which is expected.
`try_observe_self` acts the same as observe_self except that it uses `try_insert` instead of `insert` which can be used to avoid panics on the chance that the `Entity` is invalid.

Docs are fairly basic but hopefully provide the core information needed for the user. Same goes for the example. Here's hoping it's at least a good step in the right direction